+++ docs/changelog.md
@@ -4,6 +4,9 @@ nav_order: 99
 ---
 
 ## MeasureMap Changelog
+### 1.5.0
+- Refactored TaskHandler to ProcessingPipeline
+- Refactored ITaskHandler to ITaskMiddleware
 
 ### 1.4.0
 - Converted to .Net Standard library